Move internal objects from dlr.h to dlr_internal.h (#81)
